last update?

+++ JukeBox/database/migrations/2022_09_08_085638_create_song_table.php
@@ -16,7 +16,7 @@ class CreateSongTable extends Migration
         Schema::create('songs', function (Blueprint $table) {
             $table->id();
             $table->string('name', 50);
-            $table->float('time');
+            $table->integer('time');
             $table->bigInteger('rating');
             $table->timestamps();
         });
